GIP-0061: Improvements to the GIP process (WIP) (#30)
* GIP-0061: Improvements to the GIP process (WIP)

* fix: comments on the PR from Pedro, and a couple details

* chore: add flowcharts

* fix: one form to rule them all

* fix: typo

* Remove GRP Reference 0000-template

Updated the Category field to remove a reference to GRP.

* 0001-gip-process: added missing links, revised GIP stages

Added additional "Discussions-To" link to more recent forum post. Added link to new intake form wherever TODO was noted. Updated the GIP stages and descriptions.

* 0061-improved-gip-process added forum link, strawperson reference

Updated the Discussion-To portion of the form with the relevant link to the forum. Changed the Strawman reference to Strawperson to reflect the original language used.

---------

Co-authored-by: Andrew Clews <andrew@thegraph.foundation>