Update 0066-graph-horizon.md
Fixing the GIP number. If using 0066 without quotes, then it's parsed as octal (54).

strangely enough, for invalid octal numbers (like 0068), GitHub just renders as-is, which seems like it's fine, when it's not (and some parsers do throw an error)...  We should just use quotes as stated in the official GIP template.